fix tier1 and mathlib for 64bits compilation

+++ tier1/checksum_md5.cpp
@@ -285,7 +285,7 @@ void MD5Value_t::Zero()
 //-----------------------------------------------------------------------------
 bool MD5Value_t::IsZero() const
 {
-	for ( int i = 0 ; i < Q_ARRAYSIZE( bits ) ; ++i )
+	for ( unsigned long i = 0 ; i < Q_ARRAYSIZE( bits ) ; ++i )
 	{
 		if ( bits[i] != 0 )
 			return false;
